refactor: extract ContactInfo component from Contact
- Created ContactInfo component for contact details and social links
- Reduced Contact component from 574 to 462 lines (112 line reduction)
- Improved separation of concerns - contact display logic now isolated
- Added proper TypeScript interfaces for component props
- Maintained all animations and styling consistency
- Total reduction so far: 733 → 462 lines (271 line reduction)
